修复对真实 etcd 不可用的一批缺陷；重建测试防线 [minor]
六项 P0，全部在真实 etcd 3.5.17 上复现并验证修复：

- watch() 完全不可用。真实网关的 /v3/watch 是 chunked 响应，PHP http 封装会挂 dechunk
  过滤器，stream_select() 无法转换带过滤器的流 → PHP 8 抛 ValueError，回调一次都不触发。
  改为 curl 写回调 + 阻塞读双驱动，并让 watch 真正参与驱动链（此前 allow_url_fopen=0
  时即便有 curl 也是死的）。注意 curl 写回调内抛异常不会传播出 curl_exec，
  必须捕获后以短写中止再重抛，否则 watch 无法停止。
- 9 个无参方法全废：lease()->list()、memberList()、status()、defragment()、
  auth()->enable/disable() 等发送 `[]`，网关要求 `{}`，一律 400。
- keepAlive() 每次必抛而续约实际成功：网关把流式响应包在 {"result":…} 内，客户端读的是顶层。
- snapshot() 返回网关 JSON 帧而非数据库：按帧解 base64 拼接（含末尾 32 字节 sha256 摘要），
  产出 etcdutl snapshot restore 可用的 bbolt 文件。
- 认证整体失效：etcd v3 不接受 HTTP Basic，要求 /v3/auth/authenticate 换 token 并以裸
  token 发送（Bearer 前缀同样被拒）。实现换取、缓存、401 重认证，新增 AuthClient::authenticate()。
- 重试腐蚀非幂等写：5xx/读超时后重放会让 CAS 被应用两次并返回错误分支（实测"锁已归我、
  客户端却报失败"）。改为只在连接未建立时重试，只读接口额外容忍 5xx/超时。

其余：PSR-18 异常映射为 ConnectionException、options.ssl 对 curl 生效（此前三条路里两条
忽略）、3xx 不再当成功、端点故障切换不再重复挑中同一节点、uint64 ID 全程保真
（member/leader 此前被截成 PHP_INT_MAX 而指向不存在的成员）、watch 事件与 get() 返回同一
KV 形状、亚秒超时不再退化为"永不超时"、snapshot/defragment 默认 300s 且支持按调用覆盖、
instance() 传相同配置幂等、错误体嵌套对象不再 Array to string、watch 续订改为 lastRevision+1
（start_revision 是闭区间，原值会重放最后一个事件）。

测试防线（本次主要收获）：测试桩此前与真实网关在分帧（chunked）与信封（{"result":…}）
两方面都不同，导致 40+ 测试全绿而 6 个 P0 存活。重写为忠实网关桩 + 33 例回归套件
（在修复前的代码上 17 PASS / 14 FAIL，现在全过），并支持 ETCD_REAL=<host> 差分模式：
同一批用例跑真实集群并断言两者行为一致。新增 CI（单测矩阵 8.2/8.3、语法底线 8.0/8.1、
集成含真容器差分、i18n 文档）并提交 phpunit.xml.dist（此前被 gitignore，CI 跑的是默认配置）。

清理：删除 85 个手写 protobuf 桩——无任何引用、不含序列化实现，且类型是错的
（64 位用 int、枚举用 int、repeated 字段无 setter），留着会误导 gRPC 实现。

文档：13 份 README 与设计文档同步（auto 从不选 gRPC 的真相、token 认证、重试语义、
项目结构），架构图传输层与消息层重绘；补齐 id.md 此前一直缺失的代码注释翻译。

+++ src/Watch/WatchClient.php
@@ -39,7 +39,9 @@ class WatchClient
     public function watch(string $key, callable $onEvent, array $options = []): void
     {
         $rangeEnd = $options['rangeEnd'] ?? '';
-        $startRevision = $options['startRevision'] ?? 0;
+        // cast here: the transport takes an int, and a numeric string from
+        // user config would otherwise be a TypeError under strict types
+        $startRevision = (int) ($options['startRevision'] ?? 0);
         $watchOpts = [];
         if (!empty($options['prevKv'])) {
             $watchOpts['prevKv'] = true;
